Move to Climacell API version 4
The Climacell API version 3 is nearing end of life and it is no
longer possible to get an API key that works with version 3.

There are significant differences between the how the API's work
and some differences in the data returned by the V4 API.  The
biggest is that the V4 only returns data in metric units.

These changes update the node server to use the V4 API.

The V4 API no longer supports precipitation accumulation so that
has been removed from the forecast nodes.